requirements.txt
python-telegram-bot[job-queue]==21.10
configparser==6.0.0
# asyncio是Python标准库的一部分，不需要额外安装
aiosqlite==0.19.0
python-dotenv==1.0.0
psutil==5.9.8

# 搜索引擎依赖（从 tg_searcher 集成）
whoosh>=2.7.4
jieba>=0.42.1
